akash-service/requirements.txt
flask==3.0.0
google-generativeai==0.3.2
gunicorn==21.2.0